replace long() with int()
python3 doesn't have long, only int, and int works just as well in
python2.

+++ daklib/dbconn.py
@@ -748,7 +748,7 @@ class PoolFile(ORMObject):
         return os.path.basename(self.filename)
 
     def is_valid(self, filesize=-1, md5sum=None):
-        return self.filesize == long(filesize) and self.md5sum == md5sum
+        return self.filesize == int(filesize) and self.md5sum == md5sum
 
     def properties(self):
         return ['filename', 'file_id', 'filesize', 'md5sum', 'sha1sum',
